Chat_Bot for Customer Support/requirements.txt
streamlit
openai
python-telegram-bot==13.15
python-dotenv
pandas
scikit-learn
pytest
requests
kaggle
